+++ src/Assistant.Net/Data/Entities/LogSettingsEntity.cs
@@ -9,13 +9,13 @@ public class LogSettingsEntity
 {
     [Key] [Column("id")] public int Id { get; set; }
 
-    [Column("guild_id")] public decimal GuildId { get; set; }
+    [Column("guild_id")] public ulong GuildId { get; set; }
 
     [Column("log_type")] public LogType LogType { get; set; }
 
     [Column("is_enabled")] public bool IsEnabled { get; set; }
 
-    [Column("channel_id")] public decimal? ChannelId { get; set; }
+    [Column("channel_id")] public ulong? ChannelId { get; set; }
 
     [Column("delete_delay_ms")] public int DeleteDelayMs { get; set; } = 24 * 60 * 60 * 1000;
